test: move funding tx tests to ckb test module
Place funding transaction unit coverage in src/ckb/tests/fuding_tests.rs and wire module exports so test helpers remain available under cfg(test).
